add analytics and cleanup
