+++ app/src/features/rules/screens/sharedLists/components/SharedLists/analytics/constants.ts
@@ -1,0 +1,4 @@
+export const SHARED_LIST = {
+  URL_COPIED: "sharedList_url_copied",
+  DELETE_CLICKED: "shared_list_delete_clicked",
+};
