+++ app/src/features/rules/screens/sharedLists/index.ts
@@ -1,1 +1,2 @@
 export { SharedListsScreen } from "./SharedListsScreen";
+export * from "./components/SharedLists/analytics";
